backend/pyproject.toml
[build-system]
requires = ["setuptools>=42.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "data-room-api"
version = "0.1.0"
description = "Data Room API for secure document sharing"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Proprietary"}
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
dependencies = [
    "fastapi>=0.100.0",
    "uvicorn>=0.22.0",
    "sqlalchemy>=2.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-jose>=3.3.0",
    "passlib>=1.7.4",
    "python-multipart>=0.0.6",
    "email-validator>=2.0.0",
    "aiosqlite>=0.18.0",
    "asyncpg>=0.27.0",
    "loguru>=0.7.0",
    "uv>=0.1.0",
    "bcrypt>=5.0.0",
    "google-auth>=2.22.0",
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.100.0",
    "aiohttp>=3.13.2",
    "greenlet>=3.2.4",
    "python-dotenv>=1.0.0",
    "httpx>=0.24.0",
]

[tool.setuptools]
packages = {find = {where = ["."], exclude = ["uploads", "tests", "*.tests", "*.tests.*"]}}
include-package-data = true

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "httpx>=0.24.0",
    "mypy>=1.0.0",
    "ruff>=0.1.6",
    "alembic>=1.17.1",
    "greenlet>=3.2.4",
    "pytest-cov>=7.0.0",
]

[tool.black]
line-length = 88
target-version = ['py12']
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
explicit_package_bases = true
namespace_packages = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
minversion = "6.0"
addopts = "-v"
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
line-length = 88
target-version = "py312"

# Exclude directories
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Auto-fix imports
fix = true

[tool.ruff.lint]
# Enable Pyflakes `F`, pycodestyle `E`, isort `I`, flake8-bugbear `B`, and pyupgrade `UP`
select = ["E", "F", "I", "B", "UP"]
ignore = [
    "B008",  # Ignore B008 which complains about FastAPI Depends() usage
    "E501",  # Ignore line too long errors
    "E712",  # Ignore comparing against False
    "B904",  # Ignore raise from err
    "F401",  # Ignore unused imports in __init__.py files
    "UP046",  # Ignore generic class warning
]

[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "sqlalchemy"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
